Put prompts in a separate file
